broke apis with merge from encoding

+++ goftp/src/github.com/unloadedroe/common/run_length_encoding.go
@@ -10,13 +10,13 @@ func RunLengthEncoding(chunk *Chunk) *Chunk {
 
 func GetCompressedChunk(list *list.List) *Chunk {
 	compressedChunk := NewChunk()
-	newChunkContent := make([]int, list.Len()*2)
+	newChunkContent := make([]byte, list.Len()*2)
 	var index int = 0
 	for e := list.Front(); e != nil; e = e.Next() {
 		pair := e.Value.(*Pair)
-		newChunkContent[index] = pair.First().(int)
+		newChunkContent[index] = pair.First().(byte)
 		index++
-		newChunkContent[index] = pair.Second().(int)
+		newChunkContent[index] = pair.Second().(byte)
 		index++
 	}
 	compressedChunk.SetContent(&newChunkContent, index)
@@ -28,7 +28,8 @@ func GetPairs(chunk *Chunk) *list.List {
 	for oldOffset := 0; oldOffset != chunk.Len(); {
 		character := chunk.Content[oldOffset]
 		newOffset := FindBoundsForSubChunk(chunk, oldOffset)
-		pairList.PushBack(NewPair(newOffset-oldOffset, character))
+		// FIXME: cast to byte is not OK, if difference > 255 this will break.
+		pairList.PushBack(NewPair(byte(newOffset-oldOffset), character))
 		oldOffset = newOffset
 	}
 	return pairList
