+++ goftp/src/github.com/unloadedroe/common_test/run_length_encoding_test.go
@@ -7,7 +7,7 @@ import (
 
 func TestRunLengthEncoding1(t *testing.T) {
 	chunk := common.NewChunk()
-	chunk.SetContent(&[]int{'g', 'o', 'o', 'g', 'l', 'e'}, 6)
+	chunk.SetContent(&[]byte{'g', 'o', 'o', 'g', 'l', 'e'}, 6)
 	compressedChunk := common.RunLengthEncoding(chunk)
 	// (g, 103) (o, 111) (l, 108) (e, 101)
 	if compressedChunk.Content[0] != 1 && compressedChunk.Content[1] != 103 {
@@ -17,7 +17,7 @@ func TestRunLengthEncoding1(t *testing.T) {
 
 func TestRunLengthEncoding2(t *testing.T) {
 	chunk := common.NewChunk()
-	chunk.SetContent(&[]int{'g', 'o', 'o', 'g', 'l', 'e'}, 6)
+	chunk.SetContent(&[]byte{'g', 'o', 'o', 'g', 'l', 'e'}, 6)
 	compressedChunk := common.RunLengthEncoding(chunk)
 	// (g, 103) (o, 111) (l, 108) (e, 101)
 	if compressedChunk.Content[2] != 2 && compressedChunk.Content[3] != 111 {
@@ -27,7 +27,7 @@ func TestRunLengthEncoding2(t *testing.T) {
 
 func TestRunLengthEncoding3(t *testing.T) {
 	chunk := common.NewChunk()
-	chunk.SetContent(&[]int{'g', 'o', 'o', 'g', 'l', 'e'}, 6)
+	chunk.SetContent(&[]byte{'g', 'o', 'o', 'g', 'l', 'e'}, 6)
 	compressedChunk := common.RunLengthEncoding(chunk)
 	// (g, 103) (o, 111) (l, 108) (e, 101)
 	if compressedChunk.Content[4] != 1 && compressedChunk.Content[5] != 103 {
@@ -37,7 +37,7 @@ func TestRunLengthEncoding3(t *testing.T) {
 
 func TestRunLengthEncoding4(t *testing.T) {
 	chunk := common.NewChunk()
-	chunk.SetContent(&[]int{'g', 'o', 'o', 'g', 'l', 'e'}, 6)
+	chunk.SetContent(&[]byte{'g', 'o', 'o', 'g', 'l', 'e'}, 6)
 	compressedChunk := common.RunLengthEncoding(chunk)
 	// (g, 103) (o, 111) (l, 108) (e, 101)
 	if compressedChunk.Content[6] != 1 && compressedChunk.Content[7] != 108 {
@@ -47,7 +47,7 @@ func TestRunLengthEncoding4(t *testing.T) {
 
 func TestRunLengthEncoding5(t *testing.T) {
 	chunk := common.NewChunk()
-	chunk.SetContent(&[]int{'g', 'o', 'o', 'g', 'l', 'e'}, 6)
+	chunk.SetContent(&[]byte{'g', 'o', 'o', 'g', 'l', 'e'}, 6)
 	compressedChunk := common.RunLengthEncoding(chunk)
 	// (g, 103) (o, 111) (l, 108) (e, 101)
 	if compressedChunk.Content[8] != 1 && compressedChunk.Content[9] != 101 {
@@ -57,7 +57,7 @@ func TestRunLengthEncoding5(t *testing.T) {
 
 func TestGetPairs(t *testing.T) {
 	compressedChunk := common.NewChunk()
-	compressedChunk.SetContent(&[]int{'g', 'o', 'l', 'a', 'n', 'g'}, 6)
+	compressedChunk.SetContent(&[]byte{'g', 'o', 'l', 'a', 'n', 'g'}, 6)
 	listOfPairs := common.GetPairs(compressedChunk)
 	if listOfPairs.Len() != 6 {
 		t.Fatalf("size does not match!. List of pairs => %d", listOfPairs.Len())
@@ -66,7 +66,7 @@ func TestGetPairs(t *testing.T) {
 
 func TestGetPairs2(t *testing.T) {
 	compressedChunk := common.NewChunk()
-	compressedChunk.SetContent(&[]int{'g', 'o', 'o', 'g', 'l', 'e'}, 6)
+	compressedChunk.SetContent(&[]byte{'g', 'o', 'o', 'g', 'l', 'e'}, 6)
 	listOfPairs := common.GetPairs(compressedChunk)
 	if listOfPairs.Len() != 5 {
 		t.Fatalf("size does not match!. List of pairs => %d", listOfPairs.Len())
@@ -75,7 +75,7 @@ func TestGetPairs2(t *testing.T) {
 
 func TestGetPairs3(t *testing.T) {
 	compressedChunk := common.NewChunk()
-	compressedChunk.SetContent(&[]int{'g', 'o', 'o', 'o', 'l', 'e'}, 6)
+	compressedChunk.SetContent(&[]byte{'g', 'o', 'o', 'o', 'l', 'e'}, 6)
 	listOfPairs := common.GetPairs(compressedChunk)
 	if listOfPairs.Len() != 4 {
 		t.Fatalf("size does not match!. List of pairs => %d", listOfPairs.Len())
@@ -84,7 +84,7 @@ func TestGetPairs3(t *testing.T) {
 
 func TestGetPairs4(t *testing.T) {
 	compressedChunk := common.NewChunk()
-	compressedChunk.SetContent(&[]int{'g', 'o', 'o', 'o', 'o', 'e'}, 6)
+	compressedChunk.SetContent(&[]byte{'g', 'o', 'o', 'o', 'o', 'e'}, 6)
 	listOfPairs := common.GetPairs(compressedChunk)
 	if listOfPairs.Len() != 3 {
 		t.Fatalf("size does not match!. List of pairs => %d", listOfPairs.Len())
@@ -93,7 +93,7 @@ func TestGetPairs4(t *testing.T) {
 
 func TestGetPairs5(t *testing.T) {
 	compressedChunk := common.NewChunk()
-	compressedChunk.SetContent(&[]int{'g', 'o', 'o', 'o', 'o', 'o'}, 6)
+	compressedChunk.SetContent(&[]byte{'g', 'o', 'o', 'o', 'o', 'o'}, 6)
 	listOfPairs := common.GetPairs(compressedChunk)
 	if listOfPairs.Len() != 2 {
 		t.Fatalf("size does not match!. List of pairs => %d", listOfPairs.Len())
@@ -102,7 +102,7 @@ func TestGetPairs5(t *testing.T) {
 
 func TestGetPairs6(t *testing.T) {
 	compressedChunk := common.NewChunk()
-	compressedChunk.SetContent(&[]int{'o', 'o', 'o', 'o', 'o', 'o'}, 6)
+	compressedChunk.SetContent(&[]byte{'o', 'o', 'o', 'o', 'o', 'o'}, 6)
 	listOfPairs := common.GetPairs(compressedChunk)
 	if listOfPairs.Len() != 1 {
 		t.Fatalf("size does not match!. List of pairs => %d", listOfPairs.Len())
@@ -111,7 +111,7 @@ func TestGetPairs6(t *testing.T) {
 
 func TestFindBoundsForSequenceOneOffset(t *testing.T) {
 	compressedChunk := common.NewChunk()
-	compressedChunk.SetContent(&[]int{'g', 'o', 'l', 'a', 'n', 'g'}, 6)
+	compressedChunk.SetContent(&[]byte{'g', 'o', 'l', 'a', 'n', 'g'}, 6)
 	lastOffset := common.FindBoundsForSubChunk(compressedChunk, 0)
 	if lastOffset != 1 {
 		t.Fatal("offset does not match!.")
@@ -120,7 +120,7 @@ func TestFindBoundsForSequenceOneOffset(t *testing.T) {
 
 func TestFindBoundsForSequenceTwoOffset(t *testing.T) {
 	compressedChunk := common.NewChunk()
-	compressedChunk.SetContent(&[]int{'g', 'g', 'l', 'a', 'n', 'g'}, 6)
+	compressedChunk.SetContent(&[]byte{'g', 'g', 'l', 'a', 'n', 'g'}, 6)
 	lastOffset := common.FindBoundsForSubChunk(compressedChunk, 0)
 	if lastOffset != 2 {
 		t.Fatal("offset does not match!.")
@@ -129,7 +129,7 @@ func TestFindBoundsForSequenceTwoOffset(t *testing.T) {
 
 func TestFindBoundsForSequenceThreeOffset(t *testing.T) {
 	compressedChunk := common.NewChunk()
-	compressedChunk.SetContent(&[]int{'g', 'g', 'g', 'a', 'n', 'g'}, 6)
+	compressedChunk.SetContent(&[]byte{'g', 'g', 'g', 'a', 'n', 'g'}, 6)
 	lastOffset := common.FindBoundsForSubChunk(compressedChunk, 0)
 	if lastOffset != 3 {
 		t.Fatal("offset does not match!.")
@@ -138,7 +138,7 @@ func TestFindBoundsForSequenceThreeOffset(t *testing.T) {
 
 func TestFindBoundsForSequenceFourOffset(t *testing.T) {
 	compressedChunk := common.NewChunk()
-	compressedChunk.SetContent(&[]int{'g', 'g', 'g', 'g', 'n', 'g'}, 6)
+	compressedChunk.SetContent(&[]byte{'g', 'g', 'g', 'g', 'n', 'g'}, 6)
 	lastOffset := common.FindBoundsForSubChunk(compressedChunk, 0)
 	if lastOffset != 4 {
 		t.Fatal("offset does not match!.")
@@ -147,7 +147,7 @@ func TestFindBoundsForSequenceFourOffset(t *testing.T) {
 
 func TestFindBoundsForSequenceFiveOffset(t *testing.T) {
 	compressedChunk := common.NewChunk()
-	compressedChunk.SetContent(&[]int{'g', 'g', 'g', 'g', 'g', 'x'}, 6)
+	compressedChunk.SetContent(&[]byte{'g', 'g', 'g', 'g', 'g', 'x'}, 6)
 	lastOffset := common.FindBoundsForSubChunk(compressedChunk, 0)
 	if lastOffset != 5 {
 		t.Fatal("offset does not match!.")
@@ -156,7 +156,7 @@ func TestFindBoundsForSequenceFiveOffset(t *testing.T) {
 
 func TestFindBoundsForSequenceSixOffset(t *testing.T) {
 	compressedChunk := common.NewChunk()
-	compressedChunk.SetContent(&[]int{'g', 'g', 'g', 'g', 'g', 'g'}, 6)
+	compressedChunk.SetContent(&[]byte{'g', 'g', 'g', 'g', 'g', 'g'}, 6)
 	lastOffset := common.FindBoundsForSubChunk(compressedChunk, 0)
 	if lastOffset != 6 {
 		t.Fatal("offset does not match!.")
